+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Actively catching terminal resize
+- Add command aliases
 
 [0.3.0]:      https://github.com/skybet/cali/compare/v0.2.0...v0.3.0
 ## [0.3.0] - 2018-06-04